Release version 1.6
This version modified recording feature.
In concrete, when sound is recorded and WAVE file is created, ArrayBufferView is used.
Therefore, the performance was improved by using less memory than previous version.
In addition, this version corrected bugs in the case of 8 (quantization) bits.
Others, This version added transpose feature in one-shot audio.